requirements.txt
cffi==1.17.1
colorama==0.4.6
cryptography==44.0.2
pycparser==2.22
pycryptodome==3.21.0
pyperclip==1.9.0
PySide6==6.9.0
PySide6_Addons==6.9.0
PySide6_Essentials==6.9.0
shiboken6==6.9.0
